Create Appliance.js
This is the appliance 'object' (same thing as a class basically) . It is supposed to be a generalized appliance that will hold all attributes and methods that each appliance will need. This will need to be changed and updated based on the project requirements and will also probably need to be inherited from for more specific appliances that are different from the general class.